archive/backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
Authlib==1.3.2
bcrypt==4.2.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cryptography==43.0.0
cymem==2.0.8
deprecation==2.1.0
dnspython==2.6.1
dparse==0.6.4b0
ecdsa==0.19.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.112.0
filelock==3.12.4
frozenlist==1.4.1
gotrue==2.7.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
lingua-language-detector==2.0.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
nltk==3.9.1
numpy==2.0.2
outscraper==5.2.1
packaging==24.1
pip-autoremove==0.10.0
pipdeptree==2.23.1
postgrest==0.16.11
preshed==3.0.9
psutil==6.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==3.2.7
safety-schemas==0.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
supafunc==0.5.1
textblob==0.18.0.post0
thinc==8.3.2
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.1
